+++ BookBlossom/BookBlossom.Core/Interfaces/Services/IGuestService.cs
@@ -8,6 +8,6 @@ namespace BookBlossom.Core.Interfaces.Services
     {
         Task<GuestSessionResponseDTO> CreateGuestSessionAsync(string? ipAddress, string? deviceInfo);
         Task<bool> ValidateGuestSessionAsync(Guid guestId, string sessionToken);
-        Task MigrateGuestDataToUserAsync(Guid guestId, int newUserId);
+        Task MigrateGuestDataToUserAsync(Guid guestId, long newUserId);
     }
 }
